+++ src/main/kotlin/com/relogging/server/infrastructure/admin/service/AdminAuthService.kt
@@ -1,0 +1,7 @@
+package com.relogging.server.infrastructure.admin.service
+
+import com.relogging.server.infrastructure.admin.dto.AdminRequest
+
+interface AdminAuthService {
+    fun kakaoLogin(request: AdminRequest): String
+}
