+++ src/main/kotlin/com/relogging/server/infrastructure/scheduler/AdminScheduler.kt
@@ -1,0 +1,6 @@
+package com.relogging.server.infrastructure.scheduler
+
+import org.springframework.stereotype.Component
+
+@Component
+class AdminScheduler()
